Simplify playback preferences and move to :storage:preferences
