+++ app/src/main/java/de/danoeh/antennapod/dialog/SleepTimerDialog.java
@@ -27,7 +27,7 @@ import org.greenrobot.eventbus.ThreadMode;
 import java.util.Locale;
 
 import de.danoeh.antennapod.R;
-import de.danoeh.antennapod.core.preferences.SleepTimerPreferences;
+import de.danoeh.antennapod.storage.preferences.SleepTimerPreferences;
 import de.danoeh.antennapod.core.service.playback.PlaybackService;
 import de.danoeh.antennapod.ui.common.Converter;
 import de.danoeh.antennapod.core.util.playback.PlaybackController;
